# Proof of concept implementation of get_run via api (RFC).