--- a/server/fishtest/__init__.py
+++ b/server/fishtest/__init__.py
@@ -176,6 +176,7 @@ def main(global_config, **settings):
     config.add_route("api_get_elo", "/api/get_elo/{id}")
     config.add_route("api_actions", "/api/actions")
     config.add_route("api_calc_elo", "/api/calc_elo")
+    config.add_route("api_serialize_run", "/api/serialize_run/{id}")
 
     config.scan()
     return config.make_wsgi_app()