With this PR `/api/serialize_run` hosted by the main instance can be invoked by a secondary instance to receive a copy of the run from the cache.

In this way there is never an issue that the run might not be up to date.

Sadly I don't know how feasible this is. A typical serialized run is 5.5k. Is this too much?

Note: `/api/serialize_run` restricts access to `localhost` over `http`. Some form of access control is necessary but it can also be done externally.

I have tested that the PR works. The deserialized runs pass validation (for this I needed to supply the `tz_aware` codec option to `bson.decode()`).