@vdbergh 2024-05-08
Serializing of a run with 500 tasks only takes a couple of ms, but I am worried the middleware will wreak havoc with the timings. 

It would be nice to have a more direct mechanism for communicating between instances than going through nginx + middleware...