@ppigazzini 2024-05-08
> It would be nice to have a more direct mechanism for communicating between instances than going through nginx + middleware...

redis or memcached, but it's another tool to setup and maintain.