@vdbergh 2024-05-08
Another option is a named pipe. The primary instance creates it, writes the serialized run to it and communicates the name to the secondary instance. The secondary instance reads it, and then removes it. Not sure this is all possible.