@vdbergh 2024-05-08
Of course we may just as well use a real temporary file for this then.