@ppigazzini 2024-05-08
Right now PROD is able to withstand 60k cores or 100k cores with #1983.
Im fine with a clean/standard solution able to scale, not with a complex code useful only in very rarely occasion.