@ppigazzini 2024-05-08
snippet of nginx config

```nginx
server {
    listen 80;
    listen [::]:80;

    server_name dfts-0.pigazzini.it;

    location / {
        return 301 https://$host/tests;
    }

    location /api/ {
        proxy_set_header Connection "";
        proxy_set_header X-Forwarded-Proto  $scheme;
        proxy_set_header X-Forwarded-For    $proxy_add_x_forwarded_for;
        proxy_set_header X-Forwarded-Host   $host:$server_port;
        proxy_set_header X-Forwarded-Port   $server_port;
        proxy_set_header X-Country-Code     $region;

        client_max_body_size        120m;
        client_body_buffer_size     128k;
        proxy_connect_timeout       5s;
        proxy_send_timeout          30s;
        proxy_read_timeout          30s;
        proxy_max_temp_file_size    0;
        proxy_redirect              off;
        proxy_http_version          1.1;

        proxy_pass http://$backends;

        location /api/serialize_run/ {
            allow 127.0.0.1;
            deny all;
            proxy_pass http://backend_6543;
        }
    }
}


server {
    listen 443 ssl http2;
    listen [::]:443 ssl http2 ipv6only=on;

... 

    location / {
        rewrite ^/$ /tests permanent;

...

        proxy_pass http://$backends;

        location /api/serialize_run/ {
            deny all;
        }
    }
}

```