@vdbergh 2024-05-08
Well #1983 is something else. This PR is because in #1985 and #1992 the run read from the database may be missing some tasks compared to what's in the cache. This could lead to confusing behavior. This PR seems conceptually to be the most elegant way to fix that, but I am not sure if it scales...