@ppigazzini 2024-05-08
I was talking about named pipe, shared memory, not this PR.

Setup redis requires:
```bash
curl -fsSL https://packages.redis.io/gpg | sudo gpg --dearmor -o /usr/share/keyrings/redis-archive-keyring.gpg
echo "deb [signed-by=/usr/share/keyrings/redis-archive-keyring.gpg] https://packages.redis.io/deb $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/redis.list
sudo apt update
sudo apt -y install redis

sudo systemctl status redis-server
```

On python side something like this:

```python
# pip install redis

import redis
import json

r = redis.Redis(host='localhost', port=6379, db=0)

def get_run(self, r_id):
    if self.is_primary_instance():
        r_id = str(r_id)
        run_cache_item = r.get(r_id)
        if run_cache_item is not None:
            run_cache_item = json.loads(run_cache_item)
            run_cache_item["rtime"] = time.time()
            r.set(r_id, json.dumps(run_cache_item))
            return run_cache_item["run"]
        run = self.runs.find_one({"_id": ObjectId(r_id)})
        if run is not None:
            run_cache_item = {
                "rtime": time.time(),
                "ftime": time.time(),
                "run": run,
                "dirty": False,
            }
            r.set(r_id, json.dumps(run_cache_item))
        return run
    else:
        return self.runs.find_one({"_id": ObjectId(r_id)})
        
def buffer(self, run, flush):
    r_id = str(run["_id"])
    if flush:
        run_cache_item = {
            "dirty": False,
            "rtime": time.time(),
            "ftime": time.time(),
            "run": run,
        }
        r.set(r_id, json.dumps(run_cache_item))
        self.runs.replace_one({"_id": ObjectId(r_id)}, run)
    else:
        run_cache_item = r.get(r_id)
        if run_cache_item is not None:
            run_cache_item = json.loads(run_cache_item)
            ftime = run_cache_item["ftime"]
        else:
            ftime = time.time()
        run_cache_item = {
            "dirty": True,
            "rtime": time.time(),
            "ftime": ftime,
            "run": run,
        }
        r.set(r_id, json.dumps(run_cache_item))

```

But datetime and id must be converted before the dump.
```
run["start_time"] = run["start_time"].isoformat()
run["_id"] = str(run["_id"])
```