@vdbergh 2024-05-08
```
return self.runs.find_one({"_id": ObjectId(r_id)})
```
I don't understand. This still seems to return the run from the db, which is not what you want...