@vdbergh 2024-05-09
@ppigazzini Your code seems to be much more ambitious than what I am trying to do in this PR.



- In this PR I want to make it possible to ask for a secondary Fishtest instance ("the client") a copy of a run in the cache of the primary instance (e.g to have a bulletproof `get_run()`).  I was musing about how to get this copy to the client as quickly as possible (perhaps this could be done via redis).  I am not sure however if is necessary to do anything. Perhaps this PR is fast enough (I am going to optimize it a bit more by skipping the base64 encoding).

- IUC, in your code you want to replace the Fishtest cache by a redis cache. This requires a lot of thought. For example it seems to me that a naive implementation would lead to lots of unnecessary serializations/deserializations.