@vdbergh 2024-05-09
In the latest push I skip the base64 encoding. This makes the payload smaller.