@ppigazzini 2024-05-09
In nginx config we set at http level few rate limits:
- per host
- per URI
- per IP (here we have a whitelist setup)
 Also with the cookie module with a js challenge

Those rate limits should apply to this API as well. I must replicate the filtering on DEV to verify the hypothesis and test a new configuration to whitelist localhost.