@vdbergh 2024-05-09
From api.log
```
2024-05-09 17:12:01.645786+00:00 :   5.52 ms (s)  1190.12 ms (w)  https://dfts-0.pigazzini.it:443/api/upload_pgn
2024-05-09 17:14:10.054970+00:00 :   6.18 ms (s)  1191.21 ms (w)  https://dfts-0.pigazzini.it:443/api/upload_pgn
2024-05-09 17:16:19.091182+00:00 :   5.40 ms (s)  2105.73 ms (w)  https://dfts-0.pigazzini.it:443/api/upload_pgn
2024-05-09 17:18:28.921216+00:00 :   6.91 ms (s)  1098.78 ms (w)  https://dfts-0.pigazzini.it:443/api/upload_pgn
2024-05-09 17:20:32.112779+00:00 :   7.16 ms (s)  1096.62 ms (w)  https://dfts-0.pigazzini.it:443/api/upload_pgn
```