@vdbergh 2024-05-09
On PROD
```
2024-05-09 18:07:24.438469+00:00 :  21.81 ms (s)  1698.87 ms (w)  https://tests.stockfishchess.org:443/api/upload_pgn
2024-05-09 18:29:16.939441+00:00 :   6.34 ms (s)  1341.83 ms (w)  https://tests.stockfishchess.org:443/api/upload_pgn
2024-05-09 18:50:43.001719+00:00 :  56.01 ms (s)  1253.63 ms (w)  https://tests.stockfishchess.org:443/api/upload_pgn
2024-05-09 19:11:44.861195+00:00 :  21.35 ms (s)  1322.71 ms (w)  https://tests.stockfishchess.org:443/api/upload_pgn
2024-05-09 19:24:29.873764+00:00 :  37.82 ms (s)   911.89 ms (w)  https://tests.stockfishchess.org:443/api/upload_pgn
```