@ppigazzini 2024-06-28
PROD has been running with the PR for some time now.

Timings before the change:
```yaml
# logging from [28/Jun/2024:17:36:27 +0000] to [28/Jun/2024:18:51:30 +0000] ########
# duration (seconds)            : 4503
# calls in total                : 10000
# calls per second              : 2
# calls not reaching the backend: 42
# calls handled by the backend  : 9958
#                         route    calls      total    average    minimum    maximum
                         /tests      251     84.421      0.336      0.008      0.469
                   /tests/user/       99     22.129      0.224      0.035      0.275
               /api/update_task     4759     17.804      0.004      0.001      0.042
                   /api/actions     1350     12.130      0.009      0.002      0.115
                     /tests/run        8     10.708      1.339      0.490      2.408
                      /api/beat     2447      7.666      0.003      0.001      0.105
                /tests/finished       61      5.852      0.096      0.027      0.143
                /api/upload_pgn      153      3.825      0.025      0.008      0.109
               /api/active_runs      121      3.714      0.031      0.004      2.615
                       /actions        3      3.468      1.156      0.829      1.332
                  /api/get_elo/      122      3.414      0.028      0.008      0.048
                /tests/machines        9      1.931      0.215      0.178      0.250
               /tests/live_elo/       70      1.389      0.020      0.003      0.034
                   /tests/view/       50      1.307      0.026      0.013      0.054
                  /contributors        4      1.037      0.259      0.130      0.377
              /api/request_task      173      0.775      0.004      0.002      0.029
                  /tests/tasks/       11      0.593      0.054      0.022      0.190
                           /nns       19      0.430      0.023      0.009      0.121
           /api/request_version      181      0.290      0.002      0.001      0.006
                       /api/nn/       20      0.108      0.005      0.003      0.006
              /api/request_spsa       35      0.101      0.003      0.002      0.004
                 /tests/approve        2      0.040      0.020      0.013      0.027
          /contributors/monthly        1      0.034      0.034      0.034      0.034
                  /tests/delete        1      0.025      0.025      0.025      0.025
                         /login        3      0.023      0.008      0.004      0.010
               /api/failed_task        2      0.018      0.009      0.008      0.010
                              /        1      0.009      0.009      0.009      0.009
                     /sprt_calc        1      0.008      0.008      0.008      0.008
                      /workers/        1      0.003      0.003      0.003      0.003
```

Timing after the change:
```yaml
# logging from [28/Jun/2024:18:59:01 +0000] to [28/Jun/2024:20:08:07 +0000] ########
# duration (seconds)            : 4146
# calls in total                : 10000
# calls per second              : 2
# calls not reaching the backend: 37
# calls handled by the backend  : 9963
#                         route    calls      total    average    minimum    maximum
                         /tests      215     63.553      0.296      0.001      0.516
               /api/update_task     5047     19.137      0.004      0.001      0.041
                   /tests/user/       52     11.593      0.223      0.026      0.279
                   /api/actions     1135      9.883      0.009      0.002      0.050
                /api/upload_pgn      172      8.999      0.052      0.007      0.171
                      /api/beat     2369      7.177      0.003      0.001      0.031
                     /tests/run        4      5.402      1.351      0.627      2.269
                /tests/finished       45      4.634      0.103      0.026      0.133
                       /actions        7      4.419      0.631      0.080      0.947
            /api/serialize_run/      172      4.306      0.025      0.001      0.108
               /api/active_runs      106      3.344      0.032      0.004      2.456
                  /contributors        5      1.842      0.368      0.220      0.516
                /tests/machines        6      1.327      0.221      0.182      0.251
                  /api/get_elo/       54      1.321      0.024      0.008      0.045
                   /tests/view/       43      1.197      0.028      0.002      0.087
              /api/request_task      200      0.855      0.004      0.002      0.028
               /tests/live_elo/       33      0.462      0.014      0.002      0.027
           /api/request_version      228      0.366      0.002      0.001      0.004
                  /tests/tasks/        6      0.324      0.054      0.030      0.147
                      /api/pgn/        3      0.231      0.077      0.041      0.124
                           /nns       11      0.195      0.018      0.014      0.024
                              /        4      0.131      0.033      0.001      0.113
                       /api/nn/       22      0.116      0.005      0.004      0.006
              /api/request_spsa       20      0.055      0.003      0.002      0.004
                    /tests/stop        1      0.047      0.047      0.047      0.047
                 /tests/approve        2      0.027      0.013      0.013      0.014
               /api/failed_task        1      0.008      0.008      0.008      0.008

```